% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkAlignment.R
\name{checkAlignment}
\alias{checkAlignment}
\title{checkAlignment}
\usage{
checkAlignment(g, show = "both", align="b",
include=".", exclude=NULL, separateText = TRUE, rounding=4)
}
\arguments{
\item{show}{A character value indicating which hints to be shown.
Possible values are "both", "unaligned", and "aligned".}

\item{facet}{A character value indicating whether the plot showing aligned elements to be shown in facets.
Default value is "on". If "on", all the subplots (no more than 25) will be shown in facets.
If "off", all the subplots will be shown in one.
If "page", all the subplots will be shown along with prompting to the user.}

\item{align}{A numeric value indicating which hints of aligned elements to be shown.
"v" means vertical alignments, "h" means horizontal alignments and default value "b"
means both.}

\item{include}{A character value or vector indicating which kind of elements will be checked.
Default value is "." (e.g. "text", "axis", "lab")}

\item{exclude}{A character value or vector indicating which kind of elements will not be checked.
Default value is NULL. (e.g. "text", "axis", "lab")}

\item{separateText}{A logical value indicating if the elements in a textgrob will be separated.
If "TRUE", all the element in a textgrob will be regarded as different elements.
If "FALSE", all the element in a textgrob will be regarded as a unity.}

\item{rounding}{A numeric value indicating the decimal point precision used in checking.}

\item{x}{A graphic value (e.g. a ggplot)}
}
\value{
A table to show the number of each aligned element's pairs.
}
\description{
Check if the arrangement of elements follows the design principle of alignment,
that is all the elements are aligned with at least one another element of the
plot horizontally or vertically.
Mark the isolated elements in red, and aligned elements in blue.
Return a list of aligned elements' names.
}
\examples{
\dontrun{
  if (require(ggplot2)) {
    g <- ggplot(mtcars) + geom_point(aes(disp, mpg, color=factor(vs))) +
         labs(title = "test") + facet_wrap(~gear)
    checkAlignment(g, include="text", exclude=c("tag", "points"), rounding=4)
  } else {
    warning("The example requires 'ggplot2' which is not installed.")
  }
}
}
\seealso{
\code{\link[grid]{grid.ls}}
}
