% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkAlignment.R
\name{checkAlignment}
\alias{checkAlignment}
\title{Title}
\usage{
checkAlignment(
  x,
  show = "both",
  include = ".",
  exclude = NULL,
  separateText = TRUE,
  rounding
)
}
\arguments{
\item{x}{graphic}

\item{show}{both, unaligned, aligned}

\item{separateText}{define if the elements textgrob need to be separated}

\item{rounding}{}
}
\value{
draw graphics with alignment hint and return a list of aligned grobs
}
\description{
Title
}
